# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auto renginiai API",
        "description": "Auto renginių, registracijų ir autentifikacijos API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Lokali aplinka"
        }
    ],
    "paths": {
...
        "/api/auto-renginiai/{autoRenginys}/registracija": {
            "post": {
                "tags": [
                    "Registracijos"
                ],
                "summary": "Užsiregistruoti į auto renginį",
                "operationId": "eaaf1d7b3cdd1e31e174c58e1003ab22",
                "parameters": [
                    {
                        "name": "autoRenginys",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Registracija sėkminga"
                    },
                    "401": {
                        "description": "Neprisijungęs"
                    },
                    "404": {
                        "description": "Renginys nerastas"
                    },
                    "422": {
                        "description": "Negalima registruotis / validacija"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Registracijos"
                ],
                "summary": "Atšaukti registraciją į auto renginį",
                "operationId": "3042a47e494a764468183075e7c40671",
                "parameters": [
                    {
                        "name": "autoRenginys",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registracija atšaukta"
                    },
                    "401": {
                        "description": "Neprisijungęs"
                    },
                    "404": {
                        "description": "Registracijos nerasta"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}